% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydashboard_infectionvis.R
\name{epiODE}
\alias{epiODE}
\title{epiODE equation}
\usage{
epiODE(time, state, modelparams)
}
\arguments{
\item{time}{time}

\item{state}{initial states}

\item{modelparams}{parameters}
}
\value{
epiODE output
}
\description{
epiODE equation
}
\details{
epiODE
}
\examples{
init    <- c(S = 49656776, E = 0, I = 1, A = 0, Q = 0, R = 0, D = 0, Report = 1)
params  <- c(sigma1 = 1, sigma2 = 0, sigma3 = 0, sigma4 = 1, l_E = 0.1, 
             l_A = 0.5, l_Q = 0.1, gamma = 1/6, f = 0, p = 2/3, k = 1.9, alpha_Q = 1/6,
             beta = 0.33)
times   <- seq(0, 141, by = 1)
deSolve::ode(y = init, times = times, func = epiODE, parms = params)
}
