% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydashboard_infectionvis.R
\name{fluODE}
\alias{fluODE}
\title{fluODE equation}
\usage{
fluODE(time, state, params)
}
\arguments{
\item{time}{time}

\item{state}{initial states}

\item{params}{parameters}
}
\value{
fluODE output
}
\description{
fluODE equation
}
\details{
fluODE
}
\examples{
init    <- c(S = 49656776, E = 0, I = 1, A = 0, R = 0)
params  <- c(beta = 0.4726, delta = 1/2, p = 2/3, kappa = 1.9, alpha = 1/6, eta = 1/6, q = 1/2)
times   <- seq(0, 141, by = 1)
deSolve::ode(y = init, times = times, func = fluODE, parms = params)
}
