% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydashboard_infectionvis.R
\name{infectionvis}
\alias{infectionvis}
\title{'Shinydashboard' for Infection Simulation}
\usage{
infectionvis(max.filesize = 2048)
}
\arguments{
\item{max.filesize}{Maximum file size to upload (MB), Default: 2048 (2 GB)}
}
\value{
RStudio addin- 'Shinydashboard' for Infection Simulation
}
\description{
RStudio addin- 'Shinydashboard' for Infection Simulation
}
\details{
RStudio addin- 'Shinydashboard' for Infection Simulation
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[deSolve]{ode}}
}
